Code review fixes
* Stop using GenericAPIRequest
* Move GH interactions and structs to `client.go`
* Catch scopes problems
